refactor requireGroups -> requireGroup
